Create box vertex capable to draw geometry boxes

+++ engine/color.h
@@ -1,5 +1,8 @@
 #pragma once
 
+/* std includes */
+#include <cstdint> // uint8_t
+
 class color {
 public:
 	enum RGBA : unsigned {
@@ -8,16 +11,16 @@ public:
 		BLUE	= 2,
 		ALPHA	= 3
 	};
-	color(unsigned r = 0u, unsigned g = 0u, unsigned b = 0u, unsigned a = 0u) {
+	color(uint8_t r = 0u, uint8_t g = 0u, uint8_t b = 0u, uint8_t a = 0u) {
 		m_rgba[0] = r;
 		m_rgba[1] = g;
 		m_rgba[2] = b;
 		m_rgba[3] = a;
 	};
-	inline unsigned& operator[](const RGBA value) { 
+	inline uint8_t& operator[](const RGBA value) {
 		return m_rgba[value]; 
 	}
-	inline const unsigned& operator[](const RGBA value) const {
+	inline const uint8_t& operator[](const RGBA value) const {
 		return m_rgba[value];
 	}
 	color(const color& other) {
@@ -26,6 +29,7 @@ public:
 		m_rgba[RGBA::BLUE] = other[RGBA::BLUE];
 		m_rgba[RGBA::ALPHA] = other[RGBA::ALPHA];
 	}
+
 private:
-	unsigned m_rgba[4u];
+	uint8_t m_rgba[4u];
 };
